(WIP)
Two issues. I need to think of good way to get threshold setter set up with/without json combiner. Need to figure a key error message for the MAX_BG_TEMP. When that's fixed the plot functions have to be altered to depend on the right json values (depending on the stat they are using... should be easy).  Json combiner can be run on any path, so bash code is the last piece. I'll get working on that error later and then send an explanation of the changes I made. With those two issues resolved the fun begins.